Refactor the migration code for consistency
We have recently added code to migrate system connection and proxy
configs during balenaOS-to-balenaOS migrations. That code was working
correctly, but it wasn't following the same patterns as similar
previously existing code followed.

This commit refactors the new code to follow the established patterns,
which shall make takeover maintenance easier.

Being more specific, the idea is to make the new code more aligned with
the code we already had for handling WiFi networks:

1. In `migrate_info.rs` populate the `MigrateInfo` fields. This is what
   we were already doing for `wifis`.
2. In `stage1.rs` do the actual copying. Again, this is similar to what
   we did for `wifis``: this is the same place we currently already
   create the connection files for WiFi.

In summary, this is just trying to keep similar operations closer
together in the code, which hopefully will help future maintainers to
understand it.

Change-type: patch
